% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDiff.R
\name{array_preprocess}
\alias{array_preprocess}
\title{Preprocess of Microarray data}
\usage{
array_preprocess(x, missing_value = "knn", string = " /// ")
}
\arguments{
\item{x}{matrix of Microarray data, each column is a sample, 
and each row is a gene.}

\item{missing_value}{Method to  impute missing expression data,
one of "zero" and "knn".}

\item{string}{a string, sep of the gene}
}
\value{
matrix
}
\description{
Preprocess of Microarray data
}
\examples{
\donttest{
arraylist <- get_geo_array("GSE781")
arraylist <- lapply(arraylist, array_preprocess)
}
}
