% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDiff.R
\name{arrayDiff}
\alias{arrayDiff}
\title{arrayDiff}
\usage{
arrayDiff(df, group, method = "limma")
}
\arguments{
\item{df}{data.frame of the omic data}

\item{group}{a vector, group of samples.}

\item{method}{one of "limma", "ttest", "wilcox",}
}
\description{
arrayDiff
}
