% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_module.R
\name{cal_mean_module}
\alias{cal_mean_module}
\title{Title}
\usage{
cal_mean_module(geneExpress, module, result)
}
\arguments{
\item{geneExpress}{a data.frame}

\item{module}{a data.frame}

\item{result}{a string}
}
\value{
a matrix
}
\description{
Title
}
\examples{
cal_mean_module(geneExpress,module,"result.txt")
}
