% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_diff_gene.R
\name{classify_sample}
\alias{classify_sample}
\title{Title}
\usage{
classify_sample(profile_input)
}
\arguments{
\item{profile_input}{a data.frame}
}
\value{
a data.frame
}
\description{
Title
}
\examples{
profile2 <- classify_sample(profile)
}
