% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cluster_array}
\alias{cluster_array}
\title{cluster probes of Microarray data}
\usage{
cluster_array(x, clusterCutoff = 0.7)
}
\arguments{
\item{x}{matrix of Microarray data, the first is the name of the gene, 
and the others are the expression value.}

\item{clusterCutoff}{Pearson correlation threshold to cut off the hierarchical tree.}
}
\value{
data.frame
}
\description{
cluster probes of Microarray data
}
\examples{
\donttest{
arraylist <- get_geo_array("GSE781")
arraylist <- lapply(arraylist, array_preprocess)
arraylist_cluster <- lapply(arraylist, cluster_array)
}
}
