% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{Diff_limma}
\alias{Diff_limma}
\title{Diff_limma}
\usage{
Diff_limma(df, group)
}
\arguments{
\item{df}{data.frame of the omic data}

\item{group}{a vector, group of samples.}
}
\description{
Diff_limma
}
