% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{differential_limma}
\alias{differential_limma}
\title{differential_limma}
\usage{
differential_limma(df, group, adjust.method = "BH")
}
\arguments{
\item{df}{data.frame of the omic data}

\item{group}{a vector, group of samples.}

\item{adjust.method}{adjust.method.}
}
\value{
data.frame
}
\description{
differential_limma
}
\examples{
df <- matrix(runif(200), 25, 8)
df <- as.data.frame(df)
rownames(df) <- paste0("gene", 1:25)
colnames(df) <- paste0("sample", 1:8)
group <- sample(c("group1", "group2"), 8, replace = TRUE)
result <- differential_limma(df = df, group = group)
}
