% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_seq.R
\name{differential_RNA_SummarizedExperiment}
\alias{differential_RNA_SummarizedExperiment}
\title{Do difference analysis of RNA-seq data}
\usage{
differential_RNA_SummarizedExperiment(
  se,
  groupCol,
  method = "limma",
  geneLength = NULL,
  gccontent = NULL,
  filter = TRUE,
  edgeRNorm = TRUE,
  adjust.method = "BH",
  useTopconfects = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{groupCol}{group column}

\item{method}{one of "DESeq2", "edgeR" , "limma", "dearseq",
"NOISeq" and "Wilcoxon".}

\item{geneLength}{a vector of gene length.}

\item{gccontent}{a vector of gene GC content.}

\item{filter}{if TRUE, use filterByExpr to filter genes.}

\item{edgeRNorm}{if TRUE, use edgeR to do normalization for dearseq method.}

\item{adjust.method}{character string specifying the method used to
adjust p-values for multiple testing.
See \link{p.adjust} for possible values.}

\item{useTopconfects}{if TRUE, use topconfects to provide a
more biologically useful ranked gene list.}
}
\value{
data.frame
}
\description{
Do difference analysis of RNA-seq data
}
\examples{
df <- matrix(rnbinom(400, mu = 4, size = 10), 25, 16)
rownames(df) <- paste0("gene", 1:25)
colnames(df) <- paste0("sample", 1:16)
group <- sample(c("group1", "group2"), 16, replace = TRUE)

nrows <- 200; ncols <- 20
 counts <- matrix(
   runif(nrows * ncols, 1, 1e4), nrows,
   dimnames = list(paste0("cg",1:200),paste0("S",1:20))
)

colData <- S4Vectors::DataFrame(
  row.names = paste0("sample", 1:16),
  group = group
)
data <- SummarizedExperiment::SummarizedExperiment(
         assays=S4Vectors::SimpleList(counts=df),
         colData = colData)

result <- differential_RNA_SummarizedExperiment(se = data, groupCol = "group",
    filte = FALSE, method = "Wilcoxon")    
}
