% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_profile.R
\name{gene_ave}
\alias{gene_ave}
\title{Title}
\usage{
gene_ave(file1, k = 1, result = "result.txt")
}
\arguments{
\item{file1}{a data.frame}

\item{k}{a number}

\item{result}{a string for output file name}
}
\value{
a data.frame
}
\description{
Title
}
\examples{
aa <- c("Gene Symbol","MARCH1","MARC1","MARCH1","MARCH1","MARCH1")
bb <- c("GSM1629982","2.969058399","4.722410064","8.165514853","8.24243893","8.60815086")
cc <- c("GSM1629982","3.969058399","5.722410064","7.165514853","6.24243893","7.60815086")
file1 <- data.frame(aa=aa,bb=bb,cc=cc)
gene_ave(file1)
}
