% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneExpress}
\alias{geneExpress}
\title{a data.frame of gene expression data}
\format{
A data.frame with 10779 rows and 2 column
}
\usage{
geneExpress
}
\description{
It is a randomly generated expression data 
used as an example of functions in this package.
the rowname is gene symbols
the columns are gene expression values
}
\keyword{datasets}
