% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE66705_sample2}
\alias{GSE66705_sample2}
\title{a matrix of gene expression data in GEO}
\format{
A matrix with 999 rows and 3 column
}
\usage{
GSE66705_sample2
}
\description{
the first column represents the gene symbol
}
\details{
the other columns represent the expression of genes
}
\keyword{datasets}
