% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc}
\alias{hgnc}
\title{a matrix for Converting gene symbol to entrez_id or ensembl_gene_id}
\format{A matrix with 37647 rows and 5 column}
\usage{
hgnc
}
\description{
the columns represent "symbol", "locus_group", "locus_type", "entrez_id" and "ensembl_gene_id"
}
\keyword{datasets}
