% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc_file}
\alias{hgnc_file}
\title{a matrix for Converting gene symbol.}
\format{
A matrix with 43547 rows and 52 column
}
\usage{
hgnc_file
}
\description{
a matrix for Converting gene symbol.
}
\keyword{datasets}
