% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_id_conversion.R
\name{id_conversion}
\alias{id_conversion}
\title{Convert    ENSEMBL gene id to gene Symbol in TCGA}
\usage{
id_conversion(profiles, toType = "SYMBOL")
}
\arguments{
\item{profiles}{a data.frame}

\item{toType}{one of 'keytypes(org.Hs.eg.db)'}
}
\value{
a data.frame, gene symbols and their expression value
}
\description{
Convert    ENSEMBL gene id to gene Symbol in TCGA
}
\examples{
library(org.Hs.eg.db)
data(profile)
result <- id_conversion(profile)
}
