% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_id_conversion.R
\name{id_conversion_TCGA}
\alias{id_conversion_TCGA}
\title{Convert ENSEMBL gene id to gene Symbol in TCGA}
\usage{
id_conversion_TCGA(profiles, toType = "SYMBOL")
}
\arguments{
\item{profiles}{a data.frame of gene expression data, 
each column is a sample, 
and each row is a gene.}

\item{toType}{one of 'keytypes(org.Hs.eg.db)'}
}
\value{
a data.frame, gene symbols and their expression value
}
\description{
Convert ENSEMBL gene id to gene Symbol in TCGA
}
\examples{
library(org.Hs.eg.db)
data(profile)
result <- id_conversion_TCGA(profile)
}
