% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion_all.r
\name{id_conversion_vector}
\alias{id_conversion_vector}
\title{Title gene id conversion}
\usage{
id_conversion_vector(from, to, IDs)
}
\arguments{
\item{from}{one of "symbol","RefSeq_ID","Ensembl_ID","NCBI_Gene_ID","UCSC_ID","UniProt_ID"}

\item{to}{one of "symbol","RefSeq_ID","Ensembl_ID","NCBI_Gene_ID","UCSC_ID","UniProt_ID"}

\item{IDs}{the gene id which needed to convert}
}
\value{
a vector of genes
}
\description{
Title gene id conversion
}
\examples{
id_conversion_vector("symbol","Ensembl_ID",c("A2ML1","A2ML1-AS1","A4GALT","A12M1","AAAS"))
}
