% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_liver}
\alias{kegg_liver}
\title{a matrix of gene expression data in TCGA}
\format{
A matrix with 100 rows and 150 column
}
\usage{
kegg_liver
}
\description{
It is a randomly generated expression data 
used as an example of functions in this package.
the first column represents the gene symbol
}
\details{
the other columns represent the expression(count) of genes
}
\keyword{datasets}
