% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{Merge_methy_tcga}
\alias{Merge_methy_tcga}
\title{Title}
\usage{
Merge_methy_tcga(dirr)
}
\arguments{
\item{dirr}{a string for the directory of methylation data download from tcga
useing the tools gdc}
}
\value{
a matrix
}
\description{
Title
}
\examples{
mearge_result <- Merge_methy_tcga("direc")
}
