% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{methyDiff}
\alias{methyDiff}
\title{Get methylation difference gene}
\usage{
methyDiff(cpgData, sampleGroup, combineMethod = RobustRankAggreg::rhoScores)
}
\arguments{
\item{cpgData}{data.frame of cpg beta value}

\item{sampleGroup}{vector of sample group}

\item{combineMethod}{method to combine the cpg pvalues}
}
\description{
Get methylation difference gene
}
