% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{methyDiff_limma}
\alias{methyDiff_limma}
\title{methyDiff_limma}
\usage{
methyDiff_limma(df, group)
}
\arguments{
\item{df}{data.frame of the methylation data, which can be downloaded from UCSC Xena.}

\item{group}{a vector, group of samples.}
}
\description{
methyDiff_limma
}
