% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{methydifferential_SummarizedExperiment}
\alias{methydifferential_SummarizedExperiment}
\title{Get methylation difference gene}
\usage{
methydifferential_SummarizedExperiment(
  se,
  groupCol,
  combineMethod = "stouffer",
  missing_value = "knn",
  cpg2gene = NULL,
  normMethod = "PBC",
  region = "TSS1500",
  model = "gene",
  adjust.method = "BH"
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{groupCol}{group column}

\item{combineMethod}{method to combine the cpg pvalues, 
a function or one of "stouffer", "fisher" and "rhoScores".}

\item{missing_value}{Method to impute missing expression data,
one of "zero" and "knn".}

\item{cpg2gene}{data.frame to annotate cpg locus to gene}

\item{normMethod}{Method to do normalization: "PBC" or "BMIQ".}

\item{region}{region of genes, one of "Body", "TSS1500", "TSS200",
"3'UTR", "1stExon", "5'UTR", and "IGR". Only used when cpg2gene is NULL.}

\item{model}{if "cpg", step1: calculate difference cpgs;
step2: calculate difference genes.
if "gene", step1: calculate the methylation level of genes;
step2: calculate difference genes.}

\item{adjust.method}{character string specifying the method
used to adjust p-values for multiple testing.
See \link{p.adjust} for possible values.}
}
\value{
data.frame
}
\description{
Get methylation difference gene
}
\examples{
library(ChAMP)
cpgData <- matrix(runif(2000), nrow = 200, ncol = 10)
rownames(cpgData) <- paste0("cpg", seq_len(200))
colnames(cpgData) <- paste0("sample", seq_len(10))
sampleGroup <- c(rep("group1", 5), rep("group2", 5))
names(sampleGroup) <- colnames(cpgData)
cpg2gene <- data.frame(cpg = rownames(cpgData), 
    gene = rep(paste0("gene", seq_len(20)), 10))
colData <- S4Vectors::DataFrame(
    row.names = colnames(cpgData),
    group = sampleGroup
)
data <- SummarizedExperiment::SummarizedExperiment(
         assays=S4Vectors::SimpleList(counts=cpgData),
         colData = colData)
result <- methydifferential_SummarizedExperiment(se = data, 
    groupCol = "group", normMethod = NULL, 
    cpg2gene = cpg2gene)         
}
