% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module}
\alias{module}
\title{a matrix of module name, gene symbols, and the number of gene symbols}
\format{
A matrix with 176 rows and 3 column
}
\usage{
module
}
\description{
It is a randomly generated expression data 
used as an example of functions in this package.
}
\keyword{datasets}
