% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{repRemove}
\alias{repRemove}
\title{Handle the case where one id corresponds to multiple genes}
\usage{
repRemove(input_file, string)
}
\arguments{
\item{input_file}{input file, a data.frame or a matrix, 
the first column should be genes.}

\item{string}{a string,sep of the gene}
}
\value{
a data.frame, when an id corresponds to multiple genes,
the expression value is deleted
}
\description{
Handle the case where one id corresponds to multiple genes
}
\examples{
aa <- c("MARCH1 /// MMA", "MARC1", "MARCH2 /// MARCH3", 
    "MARCH3 /// MARCH4", "MARCH1")
bb <- c("2.969058399", "4.722410064", "8.165514853", 
    "8.24243893", "8.60815086")
cc <- c("3.969058399", "5.722410064", "7.165514853", 
    "6.24243893", "7.60815086")
input_file <- data.frame(aa = aa, bb = bb, cc = cc)
repRemove_result <- repRemove(input_file, " /// ")
}
