% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proce_cli.R
\name{tcga_cli_deal}
\alias{tcga_cli_deal}
\title{Combine clinical information obtained from TCGA and extract survival data}
\usage{
tcga_cli_deal(Files_dir = "your_clinical_directory")
}
\arguments{
\item{Files_dir}{a dir data}
}
\value{
a matrix, survival time and survival state in TCGA
}
\description{
Combine clinical information obtained from TCGA and extract survival data
}
\examples{
tcga_cli_deal(system.file(file.path("extdata","tcga_cli"),package="GeoTcgaData"))
}
