% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ventricle}
\alias{ventricle}
\title{a matrix of gene expression data in GEO}
\format{A matrix with 32 rows and 20 column}
\usage{
ventricle
}
\description{
the first column represents the gene symbol
}
\details{
the other columns represent the expression of genes
}
\keyword{datasets}
