% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_module.R
\name{cal_mean_module}
\alias{cal_mean_module}
\title{Find the mean value of the gene in each module}
\usage{
cal_mean_module(geneExpress, module)
}
\arguments{
\item{geneExpress}{a data.frame of gene expression data. 
Each column is a sample, and each row is a gene.}

\item{module}{a data.frame of two column. The first column is module name, 
the second column are genes in this module.}
}
\value{
a data.frame, means the mean of gene expression value in
the same module
}
\description{
Find the mean value of the gene in each module
}
\examples{
data(geneExpress)
data(module)
result <- cal_mean_module(geneExpress, module)
}
