% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_diff_gene.R
\name{diff_gene}
\alias{diff_gene}
\title{Title}
\usage{
diff_gene(profile2)
}
\arguments{
\item{profile2}{a result of classify_sample}
}
\value{
a matrix
}
\description{
Title
}
\examples{
profile2 <- classify_sample(profile)
jieguo <- diff_gene(profile2)
}
