% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{Diff_limma}
\alias{Diff_limma}
\title{Diff_limma}
\usage{
Diff_limma(df, group, adjust.method = "BH")
}
\arguments{
\item{df}{data.frame of the omic data}

\item{group}{a vector, group of samples.}

\item{adjust.method}{adjust.method.}
}
\value{
data.frame
}
\description{
Diff_limma
}
\examples{
df <- matrix(runif(200), 25, 8)
df <- as.data.frame(df)
rownames(df) <- paste0("gene", 1:25)
colnames(df) <- paste0("sample", 1:8)
group <- sample(c("group1", "group2"), 8, replace = TRUE)
result <- Diff_limma(df = df, group = group)
}
