% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_profile.R
\name{gene_ave}
\alias{gene_ave}
\title{Average the values of same genes in gene expression profile}
\usage{
gene_ave(file_gene_ave, k = 1)
}
\arguments{
\item{file_gene_ave}{a data.frame of gene expression data, 
each column is a sample, and each row is a gene.}

\item{k}{a number, indicates which is the gene column.}
}
\value{
a data.frame, the values of same genes in gene expression profile
}
\description{
Average the values of same genes in gene expression profile
}
\examples{
aa <- c("MARCH1", "MARC1", "MARCH1", "MARCH1", "MARCH1")
bb <- c(2.969058399, 4.722410064, 8.165514853, 8.24243893, 8.60815086)
cc <- c(3.969058399, 5.722410064, 7.165514853, 6.24243893, 7.60815086)
file_gene_ave <- data.frame(aa = aa, bb = bb, cc = cc)
colnames(file_gene_ave) <- c("Gene", "GSM1629982", "GSM1629983")

result <- gene_ave(file_gene_ave, 1)
}
