% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_cov}
\alias{gene_cov}
\title{a data.frame of gene length and GC content}
\format{
A data.frame with 27341 rows and 2 column
}
\usage{
gene_cov
}
\description{
the gene length and GC content data comes from 
TxDb.Hsapiens.UCSC.hg38.knownGene and
BSgenome.Hsapiens.UCSC.hg38
}
\keyword{datasets}
