% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_id_conversion.R
\name{id_conversion}
\alias{id_conversion}
\title{Convert  ENSEMBL gene id to gene Symbol in TCGA}
\usage{
id_conversion(profiles)
}
\arguments{
\item{profiles}{a data.frame}
}
\value{
a data.frame, gene symbols and their expression value
}
\description{
Convert  ENSEMBL gene id to gene Symbol in TCGA
}
\examples{
result <- id_conversion(profile)
}
