% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion_all.r
\name{id_conversion}
\alias{id_conversion}
\title{Gene id conversion}
\usage{
id_conversion(from, to, IDs, na.rm = FALSE)
}
\arguments{
\item{from}{one of `id_ava()`}

\item{to}{one of `id_ava()`}

\item{IDs}{the gene id which needed to convert}

\item{na.rm}{Whether to remove lines containing NA}
}
\value{
a vector of genes
}
\description{
Gene id conversion
}
\examples{
id_conversion(
    "symbol", "ensembl_gene_id",
    c("A2ML1", "A2ML1-AS1", "A4GALT", "A12M1", "AAAS")
)
}
