% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_id_conversion.R
\name{id_conversion}
\alias{id_conversion}
\title{Title}
\usage{
id_conversion(profile)
}
\arguments{
\item{profile}{a data.frame}
}
\value{
a data.frame
}
\description{
Title
}
\examples{
result <- id_conversion(profile)
}
