% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_methylation.R
\name{methyDiff_ucsc}
\alias{methyDiff_ucsc}
\title{Title}
\usage{
methyDiff_ucsc(methy, cpg_gene, group = NULL, missing_value = "knn")
}
\arguments{
\item{methy}{data.frame of the methylation data, which can be downloaded from UCSC Xena.}

\item{cpg_gene}{data.frame, the first coloumn is cpg id, the second coloumn is gene id.}

\item{group}{a vector of "0" and "1" for group of samples. If null, the samples were divided into two groups: disease and normal.}

\item{missing_value}{Method to  impute missing expression data, one of "zero" and "knn".}
}
\description{
Title
}
\examples{
\dontrun{
methy_file <- "TCGA.THCA.sampleMap_HumanMethylation450.gz"
methy <- data.table::fread(methy_file, sep = "\t", header = T)
library(ChAMP)
myImport <- champ.import(directory=system.file("extdata",package="ChAMPdata"))
myfilter <- champ.filter(beta=myImport$beta,pd=myImport$pd,
    detP=myImport$detP,beadcount=myImport$beadcount)
cpg_gene <- hm450.manifest.hg19[, c("probeID", "gene_HGNC")]
result <- methyDiff_ucsc(methy, cpg_gene)
}
}
