% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profile}
\alias{profile}
\title{a matrix of gene expression data in TCGA}
\format{
A matrix with 10 rows and 10 column
}
\usage{
profile
}
\description{
It is a randomly generated expression data 
used as an example of functions in this package.
the first column represents the gene symbol
}
\details{
the other columns represent the expression(FPKM) of genes
}
\keyword{datasets}
