% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proce_cli.R
\name{tcga_cli_deal}
\alias{tcga_cli_deal}
\title{Title}
\usage{
tcga_cli_deal(Files_dir1)
}
\arguments{
\item{Files_dir1}{a dir data}
}
\value{
a matrix
}
\description{
Title
}
\examples{
tcga_cli_deal("your_clinical_directory")
}
